+++ metapackages/auto-instrumentations-web/CHANGELOG.md
@@ -1,5 +1,29 @@
 # Changelog
 
+### [0.27.1](https://www.github.com/open-telemetry/opentelemetry-js-contrib/compare/auto-instrumentations-web-v0.27.0...auto-instrumentations-web-v0.27.1) (2021-12-22)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @opentelemetry/instrumentation-user-interaction bumped from ^0.27.0 to ^0.28.0
+
+## [0.27.0](https://www.github.com/open-telemetry/opentelemetry-js-contrib/compare/auto-instrumentations-web-v0.26.0...auto-instrumentations-web-v0.27.0) (2021-11-30)
+
+
+### Features
+
+* use latest instrumentation base ([#769](https://www.github.com/open-telemetry/opentelemetry-js-contrib/issues/769)) ([7aff23e](https://www.github.com/open-telemetry/opentelemetry-js-contrib/commit/7aff23ebebbe209fa3b78c2e7f513c9cd2231be4))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @opentelemetry/instrumentation-document-load bumped from ^0.26.0 to ^0.27.0
+    * @opentelemetry/instrumentation-user-interaction bumped from ^0.26.0 to ^0.27.0
+
 ## [0.26.0](https://www.github.com/open-telemetry/opentelemetry-js-contrib/compare/auto-instrumentations-web-v0.25.0...auto-instrumentations-web-v0.26.0) (2021-10-22)